Added Busniess Rules Interface and a starter implementation for Accessions A bunch of fixes for ValCombobox and getting it to work better several changes to the XML files for the views
git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@795 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39

+++ src/edu/ku/brc/specify/datamodel/RecordSet.java
@@ -142,7 +142,7 @@ public class RecordSet  implements java.io.Serializable {
 
     // Add Methods
 
-    public void addItem(final int recordId)
+    public void addItem(final Long recordId)
     {
         this.items.add(new RecordSetItem(recordId));
     }
